Fix: logger vsnprintf - warnings.

+++ src/ev/logger_v2.h
@@ -378,7 +378,7 @@ namespace ev
                 break;
             } else if ( aux > static_cast<int>(buffer_capacity_) ) {
                 // ... realloc buffer ...
-                if ( true == EnsureBufferCapacity(static_cast<size_t>(aux + sizeof(char))) ) {
+                if ( true == EnsureBufferCapacity(static_cast<size_t>((unsigned long)aux + sizeof(char))) ) {
                     // ... last attempt to write to buffer ...
                     continue;
                 } else {
